backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
Pillow==9.5.0
httpx==0.27.0
#Uncomment below when using real Gemini API:
google-generativeai==0.5.4
